# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server for Resource Extraction WebApp",
        "title": "Resource Extraction",
        "contact": {
            "name": "API Support",
            "url": "https://vk.com/bmstu_schedule",
            "email": "bitop@spatecon.ru"
        },
        "license": {
            "name": "AS IS (NO WARRANTY)"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/home": {
            "get": {
                "description": "Загружает главную страницу с ресурсами или выполняет поиск ресурсов по названию.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Загрузка главной страницы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название ресурса для поиска",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/home/add_resource": {
            "post": {
                "description": "Добавляет новый ресурс с соответсвующими параметрами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Добавление нового ресурса",
                "parameters": [
                    {
                        "description": "Ресурс",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.AddResRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/home/delete_report/{title}": {
            "post": {
                "description": "Логическое удаление отчета из таблицы отчетов и физическое от таблицы ММ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Удаление отчета",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID отчета",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/home/delete_resource/{title}": {
            "post": {
                "description": "Логически удаляет ресурс (меняет статус)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Удаление ресурса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название ресурса",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/home/get_report/{title}": {
            "get": {
                "description": "Получение отчета по его айди",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Получение отчета по его айди",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID отчета",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/home/get_report/{title}/change_status": {
            "put": {
                "description": "Изменение статуса у отчета с ограничениями",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Изменить статус у отчета",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID отчета",
                        "name": "title",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Кто меняет / на какой статус",
                        "name": "change",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.ChangeStatusRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/home/{title}": {
            "get": {
                "description": "Загружает страницу с определенным ресурсом и информацию о нем",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Загрузка страницы ресурса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название ресурса",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/home/{title}/add_monthly_prod": {
            "post": {
                "description": "Если записей о добыче ресурса еще нет, то изменяет эту запись. Если же информация о добыче за какие-то месяцы уже есть, то создает новую запись",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Добавление информации о месячной добычи",
                "parameters": [
                    {
                        "description": "Месячная доыбча",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.AddMonthlyProd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/home/{title}/add_report": {
            "post": {
                "description": "Добавление отчета по добыче по какому-то ресурса (по месту, в котором он добывается)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Добавление отчета о добыче",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название ресурса",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.AddMonthlyProd": {
            "type": "object",
            "properties": {
                "month": {
                    "type": "string"
                },
                "monthlyProd": {
                    "type": "number"
                }
            }
        },
        "ds.AddResRequestBody": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "place": {
                    "type": "string"
                },
                "resourceName": {
                    "type": "string"
                }
            }
        },
        "ds.ChangeStatusRequestBody": {
            "type": "object",
            "properties": {
                "new_status": {
                    "type": "string"
                },
                "who": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}